app/requirements.txt
streamlit
python-dotenv
boto3
openai